Make coin fountain with particles
Instead of spawning a bunch of coin objects that collide with each
other, make a particle system. Once I manipulated the vertexes directly
in blender of my coin object I was able to add these coins to the mesh!
Only con is that when they land on the surface they won't turn, but that
could be a feature instead of a bug depending on how the effect looks
when we use this in AR.