Test Clip Distance & Palm Detections
Found that palm detection is supported in Lightship, so definitely want to try that api out. Otherwise since my coin fountains aren't showing I'm 1. creating a few objects in the scene to show that I'm can see the coin fountains and 2. expanding clip distance in case that's impacting our scenario.