Palm And Bundle
Clip experiment didn't work, but before I had seen that cubes were being spawned where I clicked. Therefore to determine if it's a particle issue or something else I've bundled the cube and the particle system together.

In the meantime I hadn't enabled the palm system so did that too.