Clicking 4 Points Generates Mesh
On Android Lightship's mesh is unstable, even with adjustments I made.
Therefore instead of using the mesh directly I'm going to make a mesh
based on points I specify.

This commit allows you to click on 4 points and generate a mesh between
those points. Cubes are generated at the points of contact to confirm
the user's input has been percieved.