Use Raycasting to Create a board
Instead of making a gameboard that is always on, I want to make a
gameboard that appears when we interact with the app. For right now that
means tapping the screen. The BoardGenerator class is replaced with the
GameGenerator which uses iterative raycasts to establish a flat area of the
scene. Right now we show the farthest, closest, left, and right edges of the
area with cubes since it's a simple debugging tool.